feat: user delete confirm implementation
