+++ src/components/DeleteItem.dialog.tsx
@@ -38,7 +38,7 @@ const DeleteItemDialog = ({
 
 					<AlertDialogBody>
 						{subtitle ??
-							'Are you sure? You can&apos;t undo this action afterwards.'}
+							`Are you sure? You can't undo this action afterwards.`}
 					</AlertDialogBody>
 
 					<AlertDialogFooter>
